+++ one_fm/api/mobile/Leave_application.py
@@ -8,7 +8,7 @@ import collections
 import base64, json
 from one_fm.api.v1.utils import response, validate_date
 from frappe.utils import cint, cstr, getdate
-from one_fm.api.v1.roster import get_current_shift
+from one_fm.utils import get_current_shift
 from one_fm.api.tasks import get_action_user
 from one_fm.api.api import push_notification_rest_api_for_leave_application
 from one_fm.processor import sendemail
