+++ one_fm/api/v2/face_recognition.py
@@ -1,8 +1,8 @@
 from datetime import timedelta, datetime
 import frappe, ast, base64, time, grpc, json, random
 from frappe import _
-from one_fm.one_fm.page.face_recognition.utils import check_existing, update_onboarding_employee, late_checkin_checker
-from one_fm.api.v1.roster import get_current_shift
+from one_fm.one_fm.page.face_recognition.utils import update_onboarding_employee, late_checkin_checker
+from one_fm.utils import get_current_shift, check_existing
 from one_fm.api.v1.utils import response
 from one_fm.api.v2.zenquotes import fetch_quote
