+++ one_fm/one_fm/page/face_recognition/face_recognition.py
@@ -13,9 +13,9 @@ from json import JSONEncoder
 import json
 # from imutils import face_utils, paths
 from one_fm.api.doc_events import haversine
-from one_fm.api.v1.roster import get_current_shift
+from one_fm.utils import get_current_shift, check_existing
 
-from one_fm.one_fm.page.face_recognition.utils import check_existing, update_onboarding_employee, late_checkin_checker
+from one_fm.one_fm.page.face_recognition.utils import update_onboarding_employee, late_checkin_checker
 from one_fm.api.v2.zenquotes import fetch_quote
 
 # setup channel for face recognition
